+++ tests/Unit/PublisherTest.php
@@ -8,6 +8,7 @@ use Convenia\Pigeon\Tests\TestCase;
 use PhpAmqpLib\Channel\AMQPChannel;
 use PhpAmqpLib\Message\AMQPMessage;
 use Convenia\Pigeon\Publisher\Publisher;
+use PhpAmqpLib\Wire\AMQPTable;
 
 class PublisherTest extends TestCase
 {
@@ -159,7 +160,9 @@ class PublisherTest extends TestCase
         // assert
         $this->channel->shouldReceive('basic_publish')->with(
             Mockery::on(function (AMQPMessage $arg) use ($headers) {
-                return Arr::dot($headers) === $arg->get_properties()['application_headers'];
+                $app_headers = $arg->get('application_headers');
+                return ($app_headers instanceof AMQPTable)
+                    && ($headers === $app_headers->getNativeData());
             }),
             $exchange,
             null