+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ## Fixed
 - Auto declare exchange with `routing`
+- Fix `application_headers` to `AMQPTable`
 
 ## [1.1.0]
 ### Added